go channel

+++ controllers/match/match.go
@@ -138,7 +138,7 @@ func ActiveUserUpper(ctx *quick_match.QuickMatchContext) {
 	before24HourTime := nowTime - offsetTime
 	for i, user := range ctx.UserList {
 		if user.UserCache.LastUpdateTime >= before24HourTime {
-			var addRate = float32((user.UserCache.LastUpdateTime - before24HourTime) / offsetTime) * upperRate
+			var addRate = float32(user.UserCache.LastUpdateTime - before24HourTime) / float32(offsetTime) * upperRate
 			ctx.UserList[i].Score = ctx.UserList[i].Score * (1.0 + addRate)
 			log.Infof("ActiveUserUpper before:%d, last:%d, add:%.3f old:%.3f new:%.3f", 
 			before24HourTime, user.UserCache.LastUpdateTime, addRate, ctx.UserList[i].AlgoScore, ctx.UserList[i].Score)
